requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
<<<<<<< HEAD
=======
plotly==5.24.
>>>>>>> 86ab4975cf71cee1207670d52a14b29e632e7352
plotly>=5.15.0
scipy>=1.11.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
xlrd>=2.0.0
pyarrow>=12.0.0
fastparquet>=0.8.0
statsmodels>=0.14.0
seaborn>=0.12.0
matplotlib>=3.7.0
umap-learn>=0.5.3
hdbscan>=0.8.29
fpdf2>=2.7.0
Pillow>=10.0.0
kaleido>=0.2.1
<<<<<<< HEAD
reportlab>=4.0.0
prophet>=1.1.5
python-docx>=0.8.11
python-pptx>=0.6.21
=======
<<<<<<<< HEAD:requirements_premium.txt
reportlab>=4.0.0
prophet>=1.1.5
python-docx>=0.8.11
python-pptx>=0.6.21
========
>>>>>>>> 86ab4975cf71cee1207670d52a14b29e632e7352:requirements.txt
>>>>>>> 86ab4975cf71cee1207670d52a14b29e632e7352
